moved shops related code to shops folder
